requirements.txt
Flask==3.0.0
supabase==2.10.0
python-dotenv==1.0.0
Werkzeug==3.0.1
